Refactor to transcend some design flaws
With the necessity to start allowing for multiple Listen addresses, a
number of design flaws surfaced. These flaws mainly revolved around
scoping and parse order issues that were easily addressed when
relevant code was refactored.

Some major changes made in this commit are the changing of the
apache::namevirtualhost class into a defined type that is called
automatically as needed. Similarly, the apache::listen class was created
that behaves nearly identical to the apache::namevirtualhost define.

Both of these defines use concat to put snippets of the Listen and
NameVirtualHost directives into a common file that is loaded once in
httpd.conf.

This change required some reworking and consequently some major
restructuring of code and spec tests. In addition, it adds the
dependency on the concat module, 0.2.0 or greater.