(fix) remove fail and comment on the anchor pattern usage
we remove the fail statement because sometimes the parse order is such
that the apache class may not have been parsed yet, but the anchor
pattern appears to ensure it's values are defined before it gets to
mod_ssl. more testing needed to confirm it works, but for now we know
that this gets in the way.