remove ability to override confd_dir in apache class
the puppet parser does not parse manifests in a reasonanble nor
consistent order. As a result, in certain cases classes relying on
values in the main apache class (ie. $::apache::confd_dir) will be
returned an empty string because the apache class has not been evaluated
yet.

Because resource ordering does not affect parser ordering, which at this
time is unclear as to how it exactly functions, I'm removing this
ability since it does not work consistently and can cause subtle issues
that are difficult to track down and even more difficult to resolve.

The puppet parser, in this case, sucks.

A bug exists for this, see the following link for more information:
https://tickets.puppetlabs.com/browse/PUP-1417